myparliament/requirements.txt
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
Pillow==7.2.0
Scrapy==2.3.0
scrapy-djangoitem==1.1.1
